- If we are unable to obtain the latest version of a package using the http api of nuget, instead of defaulting to version 1.0.0, we return null, which will resort in obtaining the latest package from Nuget as a failsafe.
- Updated nuget version from 2.7.2 to 2.8.1

Removed the need to use the nuget url api to get the latest installed version.

Created a caching mechanism to store the version of the Nuget packages within the solution, so projects within the same solution will be able to install the versions in a performant way as opposed to getting the latest version of the package all the time, which includes performance cost.